🚚 rename: 이벤트/http 서비스 분리
