+++ src/main/java/com/mulmeong/reward/point/application/MemberPointEventService.java
@@ -1,0 +1,8 @@
+package com.mulmeong.reward.point.application;
+
+import com.mulmeong.event.member.MemberCreateEvent;
+
+public interface MemberPointEventService {
+
+    void createMemberPointDocument(MemberCreateEvent memberCreateEvent);
+}
