+++ src/main/java/com/mulmeong/reward/point/application/MemberPointHttpService.java
@@ -1,0 +1,8 @@
+package com.mulmeong.reward.point.application;
+
+import com.mulmeong.reward.point.dto.out.MemberPointDto;
+
+public interface MemberPointHttpService {
+
+    MemberPointDto getMemberPoint(String memberUuid);
+}
