+++ src/main/java/com/mulmeong/reward/point/infrastructure/PointKafkaConsumer.java
@@ -1,7 +1,7 @@
 package com.mulmeong.reward.point.infrastructure;
 
 import com.mulmeong.event.member.MemberCreateEvent;
-import com.mulmeong.reward.point.application.MemberPointService;
+import com.mulmeong.reward.point.application.MemberPointEventService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
@@ -12,12 +12,12 @@ import org.springframework.stereotype.Service;
 @Service
 public class PointKafkaConsumer {
 
-    private final MemberPointService memberPointService;
+    private final MemberPointEventService memberPointEventService;
 
     @KafkaListener(topics = "${event.member.pub.topics.member-create.name}",
             containerFactory = "memberCreateEventListener")
     public void handleMemberCreatedEvent(MemberCreateEvent event) {
         log.info("Consumed 회원 생성 이벤트 : {}", event);
-        memberPointService.createMemberPointDocument(event);
+        memberPointEventService.createMemberPointDocument(event);
     }
 }
